added some userwarnings an API design
